// src/redux/Types/quizTypes.ts
export enum QuizTypes {
    FETCH_ALL_QUIZES = 'FETCH_ALL_QUIZES',
    FETCH_USER_QUIZES = 'FETCH_USER_QUIZES',
    SET_FORM_COUNTER = 'SET_FORM_COUNTER',
    SET_FORM_INPUT_TITLE = 'SET_FORM_INPUT_TITLE',
    SET_FORM_COLOR = 'SET_FORM_COLOR',
    SET_FORM_ICON_NAME = 'SET_FORM_ICON_NAME',

    SET_FORM_RADIO = 'SET_FORM_RADIO',
    SET_FORM_QUESTION = 'SET_FORM_QUESTION',
    SET_FORM_ANSWER_A = 'SET_FORM_ANSWER_A',
    SET_FORM_ANSWER_B = 'SET_FORM_ANSWER_B',
    SET_FORM_ANSWER_C = 'SET_FORM_ANSWER_C',
    SET_FORM_ANSWER_D = 'SET_FORM_ANSWER_D'
}